Reorder routes and controllers to new curse_of_mirra folder
