+++ apps/gateway/lib/gateway/router.ex
@@ -9,10 +9,12 @@ defmodule Gateway.Router do
     pipe_through :api
   end
 
-  scope "/curse", Gateway.Curse do
+  scope "/curse", Gateway.Controllers.CurseOfMirra do
     pipe_through :api
 
-    put "/users/:user_id/currency", Controllers.Users.CurrencyController, :modify_currency
+    put "/users/:user_id/currency", CurrencyController, :modify_currency
+    get "/users/:user_id/claim_daily_reward", UserController, :claim_daily_reward
+    get "/users/:user_id/get_daily_reward_status", UserController, :get_daily_reward_status
   end
 
   scope "/arena", Gateway.Controllers.Arena do
@@ -27,8 +29,6 @@ defmodule Gateway.Router do
     get "/auth/:provider/token/:token_id", Controllers.AuthController, :validate_token
 
     put "/users/:user_id", Controllers.UserController, :update
-    get "/users/:user_id/claim_daily_reward", Controllers.UserController, :claim_daily_reward
-    get "/users/:user_id/get_daily_reward_status", Controllers.UserController, :get_daily_reward_status
   end
 
   # Other scopes may use custom stacks.
